+++ core/core-domain/src/main/java/com/barlow/core/domain/billpost/BillPostDetailQuery.java
@@ -1,4 +1,4 @@
-package com.barlow.core.domain.recentbill;
+package com.barlow.core.domain.billpost;
 
 public record BillPostDetailQuery(String billId) {
 }
