+++ storage/db-core/src/main/java/com/barlow/storage/db/core/BillProposerJpaEntity.java
@@ -1,6 +1,6 @@
 package com.barlow.storage.db.core;
 
-import com.barlow.core.domain.recentbill.BillProposer;
+import com.barlow.core.domain.billpost.BillProposer;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
